travel_planner_guide/agent.py
from google.adk.agents import Agent
from travel_planner_guide.supporting_agents import travel_inspiration_agent


LLM="gemini-2.5-flash"


root_agent = Agent(
    model = LLM,
    name= "travel_planner_main_agent",
    description= "A helpful travel planning assistant that helps users plan their trips by providing informations and suggestions acording to their preferences.",
    instruction = """
            - You are an exclusive travel concierge agent
            - You help users to discover their dream holiday destinations and plan their vacation.
            - Use the inspiration_agent to get the best destination, news, places nearby e.g hotels, cafes, etc near attractions and points of interest for the user.
            - You cannot use any tool directly.
            """,
    sub_agents = [travel_inspiration_agent]
    
)
